API/requirements.txt
fastapi
uvicorn
pillow
opencv-python
numpy
PyMuPDF
google-generativeai
python-multipart
